mandos step reorg
