+++ elrond-wasm-debug/src/mandos_system/model/step/mod.rs
@@ -1,5 +1,19 @@
+mod step_check_state;
+mod step_dump;
 mod step_enum;
+mod step_sc_call;
+mod step_sc_deploy;
+mod step_sc_query;
 mod step_set_from_proxy;
-mod step_sugar;
+mod step_set_state;
+mod step_transfer;
 
+pub use step_check_state::*;
+pub use step_dump::*;
 pub use step_enum::*;
+pub use step_sc_call::*;
+pub use step_sc_deploy::*;
+pub use step_sc_query::*;
+pub use step_set_from_proxy::*;
+pub use step_set_state::*;
+pub use step_transfer::*;
