+++ elrond-wasm-debug/src/mandos_system/model/step/step_dump.rs
@@ -1,0 +1,4 @@
+#[derive(Debug, Default)]
+pub struct DumpStateStep {
+    pub comment: Option<String>,
+}
