+++ elrond-wasm-debug/src/mandos_system/model/step/step_sc_call.rs
@@ -1,0 +1,90 @@
+use crate::mandos_system::model::{
+    AddressValue, BigUintValue, BytesValue, TxCall, TxESDT, TxExpect, U64Value,
+};
+
+#[derive(Debug, Default)]
+pub struct ScCallStep {
+    pub tx_id: String,
+    pub comment: Option<String>,
+    pub tx: Box<TxCall>,
+    pub expect: Option<TxExpect>,
+}
+
+impl ScCallStep {
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    pub fn from<A>(mut self, address: A) -> Self
+    where
+        AddressValue: From<A>,
+    {
+        self.tx.from = AddressValue::from(address);
+        self
+    }
+
+    pub fn to<A>(mut self, address: A) -> Self
+    where
+        AddressValue: From<A>,
+    {
+        self.tx.to = AddressValue::from(address);
+        self
+    }
+
+    pub fn egld_value<A>(mut self, amount: A) -> Self
+    where
+        BigUintValue: From<A>,
+    {
+        if !self.tx.esdt_value.is_empty() {
+            panic!("Cannot transfer both EGLD and ESDT");
+        }
+
+        self.tx.egld_value = BigUintValue::from(amount);
+        self
+    }
+
+    pub fn esdt_transfer<T, N, A>(mut self, token_id: T, token_nonce: N, amount: A) -> Self
+    where
+        BytesValue: From<T>,
+        U64Value: From<N>,
+        BigUintValue: From<A>,
+    {
+        if self.tx.egld_value.value > 0u32.into() {
+            panic!("Cannot transfer both EGLD and ESDT");
+        }
+
+        self.tx.esdt_value.push(TxESDT {
+            esdt_token_identifier: BytesValue::from(token_id),
+            nonce: U64Value::from(token_nonce),
+            esdt_value: BigUintValue::from(amount),
+        });
+
+        self
+    }
+
+    pub fn function(mut self, expr: &str) -> Self {
+        self.tx.function = expr.to_string();
+        self
+    }
+
+    pub fn argument<A>(mut self, expr: A) -> Self
+    where
+        BytesValue: From<A>,
+    {
+        self.tx.arguments.push(BytesValue::from(expr));
+        self
+    }
+
+    pub fn gas_limit<V>(mut self, value: V) -> Self
+    where
+        U64Value: From<V>,
+    {
+        self.tx.gas_limit = U64Value::from(value);
+        self
+    }
+
+    pub fn expect(mut self, expect: TxExpect) -> Self {
+        self.expect = Some(expect);
+        self
+    }
+}
