+++ elrond-wasm-debug/src/mandos_system/model/step/step_sc_query.rs
@@ -1,0 +1,38 @@
+use crate::mandos_system::model::{AddressValue, BytesValue, TxExpect, TxQuery};
+
+#[derive(Debug, Default)]
+pub struct ScQueryStep {
+    pub tx_id: String,
+    pub comment: Option<String>,
+    pub tx: Box<TxQuery>,
+    pub expect: Option<TxExpect>,
+}
+
+impl ScQueryStep {
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    pub fn to<A>(mut self, address: A) -> Self
+    where
+        AddressValue: From<A>,
+    {
+        self.tx.to = AddressValue::from(address);
+        self
+    }
+
+    pub fn function(mut self, expr: &str) -> Self {
+        self.tx.function = expr.to_string();
+        self
+    }
+
+    pub fn argument(mut self, expr: &str) -> Self {
+        self.tx.arguments.push(BytesValue::from(expr));
+        self
+    }
+
+    pub fn expect(mut self, expect: TxExpect) -> Self {
+        self.expect = Some(expect);
+        self
+    }
+}
